% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_words_summary.R
\name{all_words_summary}
\alias{all_words_summary}
\title{A high level summary about all the words}
\usage{
all_words_summary(data, column = "word")
}
\arguments{
\item{data}{A data frame containing a column of words}

\item{column}{A character column containing a list of words}
}
\value{
A tibble
}
\description{
A high level summary about all the words
}
\examples{
example_words <- data.frame(word_column = c("alfa", "bravo", "charlie", "delta"))
all_words_summary(data = example_words, column = "word_column")

}
