% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starting_letter_plot.R
\name{starting_letter_plot}
\alias{starting_letter_plot}
\title{See distribution of starting letters for words}
\usage{
starting_letter_plot(df, column = "word")
}
\arguments{
\item{df}{A data frame containing a column of words}

\item{column}{A character column containing a list of words}
}
\value{
A starting letter distribution plot
}
\description{
See distribution of starting letters for words
}
\examples{
dummy_words <- data.frame(word = c("alfa", "bravo", "charlie", "delta"))
starting_letter_plot(dummy_words)
}
