% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_words_plot.R
\name{top_words_plot}
\alias{top_words_plot}
\title{Plot the most common words in a column of a dataframe.}
\usage{
top_words_plot(data, column = "word", top_n = 6)
}
\arguments{
\item{data}{A data frame containing a column of words}

\item{column}{A character column containing a list of words}

\item{top_n}{An integer that specifies how many words you want in the plot}
}
\value{
A ggplot2 bar chart of the most common words
}
\description{
Plot the most common words in a column of a dataframe.
}
\examples{
example_words <- data.frame(word_column = c("alfa", "bravo", "charlie", "delta", "alfa"))
top_words_plot(example_words, column = "word_column", top_n = 4)

}
