% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_characteristics.R
\name{word_characteristics}
\alias{word_characteristics}
\title{An in-depth summary of each word}
\usage{
word_characteristics(data, column = "word")
}
\arguments{
\item{data}{A data frame containing a column of words}

\item{column}{A character column containing a list of words}
}
\value{
A data frame
}
\description{
An in-depth summary of each word
}
\examples{
example_words <- data.frame(word_column = c("alfa", "bravo", "charlie", "delta", "civic"))
word_characteristics(data = example_words, column = "word_column")

}
