% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_length_plot.R
\name{word_length_plot}
\alias{word_length_plot}
\title{Plot word length distribution}
\usage{
word_length_plot(data, column = "word")
}
\arguments{
\item{data}{A data frame containing a column of words}

\item{column}{A character column containing a list of words}
}
\value{
A ggplot2 bar chart showing the distribution of word lengths
}
\description{
Plot word length distribution
}
\examples{
example_words <- data.frame(my_words = c("alfa", "bravo", "charlie", "delta"))
word_length_plot(data = example_words, column = "my_words")

}
