% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_words_summary.R
\name{all_words_summary}
\alias{all_words_summary}
\title{A high level summary about all the words}
\usage{
all_words_summary(df, column = "word")
}
\arguments{
\item{df}{A data frame containing a column of words}

\item{column}{A character column containing a list of words}
}
\value{
A tibble
}
\description{
A high level summary about all the words
}
\examples{
dummy_words <- data.frame(word = c("alfa", "bravo", "charlie", "delta"))
all_words_summary(dummy_words)
}
