% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_words_plot.R
\name{top_words_plot}
\alias{top_words_plot}
\title{Top words bar plot}
\usage{
top_words_plot(df, column = "word", top_n = 6)
}
\arguments{
\item{df}{A data frame containing a column of words}

\item{column}{A character column containing a list of words}

\item{top_n}{An integer that specifies how many words you want in the plot}
}
\value{
A ggplot2 bar chart of the most common words
}
\description{
See most common words in a column of a dataframe.
}
\examples{
dummy_words <- data.frame(word = c("alfa", "bravo", "charlie", "delta"))
top_words_plot(dummy_words)
}
