% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_characteristics.R
\name{word_characteristics}
\alias{word_characteristics}
\title{An in-depth summary of each word}
\usage{
word_characteristics(df, column = "word")
}
\arguments{
\item{df}{A data frame containing a column of words}

\item{column}{A character column containing a list of words}
}
\value{
A data frame
}
\description{
An in-depth summary of each word
}
\examples{
dummy_words <- data.frame(word = c("alfa", "bravo", "charlie", "delta"))
word_characteristics(dummy_words)
}
