% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_length_plot.R
\name{word_length_plot}
\alias{word_length_plot}
\title{See word length distribution}
\usage{
word_length_plot(df, column = "word")
}
\arguments{
\item{df}{A data frame containing a column of words}

\item{column}{A character column containing a list of words}
}
\value{
A word length plot
}
\description{
See word length distribution
}
\examples{
dummy_words <- data.frame(word = c("alfa", "bravo", "charlie", "delta"))
word_length_plot(dummy_words)
}
